Rename HiSPARC event message files
